feat: add ChatPanel component
