+++ view/pages/Questions/VibeChat.tsx
@@ -1,17 +1,12 @@
-import { Reference, useReactiveVar } from '@apollo/client';
+import { Reference } from '@apollo/client';
 import { Typography } from '@mui/material';
-import { UIEvent, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import MessageFeed from '../../components/Chat/MessageFeed';
-import MessageForm from '../../components/Chat/MessageForm';
+import ChatPanel from '../../components/Chat/ChatPanel';
 import ProgressBar from '../../components/Shared/ProgressBar';
-import { scrollDirectionVar } from '../../graphql/cache';
 import { useVibeChatQuery } from '../../graphql/chat/queries/gen/VibeChat.gen';
 import { useNewMessageSubscription } from '../../graphql/chat/subscriptions/gen/NewMessage.gen';
 
 const VibeChat = () => {
-  const [feedScrollPosition, setFeedScrollPosition] = useState(0);
-  const scrollPosition = useReactiveVar(scrollDirectionVar);
   const { data, loading, error } = useVibeChatQuery();
 
   useNewMessageSubscription({
@@ -46,28 +41,6 @@ const VibeChat = () => {
   });
 
   const { t } = useTranslation();
-  const feedRef = useRef<HTMLDivElement>(null);
-
-  const handleScroll = (e: UIEvent<HTMLDivElement>) => {
-    const target = e.target as HTMLDivElement;
-    setFeedScrollPosition(target.scrollTop);
-  };
-
-  const handleImageLoad = () => {
-    if (
-      feedRef.current &&
-      feedScrollPosition > -100 &&
-      scrollPosition !== 'up'
-    ) {
-      feedRef.current.scrollIntoView();
-    }
-  };
-
-  const handleSubmit = () => {
-    if (feedRef.current) {
-      feedRef.current.scrollIntoView();
-    }
-  };
 
   if (error) {
     return <Typography>{t('errors.somethingWentWrong')}</Typography>;
@@ -82,22 +55,11 @@ const VibeChat = () => {
   }
 
   return (
-    <>
-      <MessageFeed
-        feedRef={feedRef}
-        messages={data.vibeChat.messages}
-        onImageLoad={handleImageLoad}
-        onScroll={handleScroll}
-      />
-
-      {data && (
-        <MessageForm
-          conversationId={data.vibeChat.id}
-          onSubmit={handleSubmit}
-          vibeChat
-        />
-      )}
-    </>
+    <ChatPanel
+      conversationId={data.vibeChat.id}
+      messages={data.vibeChat.messages}
+      vibeChat
+    />
   );
 };
